+++ STARBUCKS_iOS/STARBUCKS_iOS/Presentation/Entity/DrinkDetailEntity.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct DrinkDetailEntity {
-    let myMenuId: Int64
+    let myMenuId: Int
     let isHot: Bool
 
     let hotMenuKr: String
